requirements.txt
flask
flask_jwt_extended
kafka-python>=2.0.2
pymongo
apscheduler